+++ ChangeLog.md
@@ -20,6 +20,12 @@
 - **BREAKING CHANGE:** Gopher configuration is now loaded from `gopher.yaml`
   rather than `configuration.json`.
   ([#448](https://github.com/davep/rogallo/pull/448))
+- Command line history is now stored in the application's state directory.
+  ([#452](https://github.com/davep/rogallo/pull/452))
+- Location history is now stored in the application's state directory.
+  ([#452](https://github.com/davep/rogallo/pull/452))
+- Navigation history is now stored in the application's state directory.
+  ([#452](https://github.com/davep/rogallo/pull/452))
 
 ## v2.4.0